moving utils method to a better named file
